scraping/scrape_flights.py
# scrape_flights.py
"""
Scraping logic for airline booking data from web sources.
"""

import requests
from bs4 import BeautifulSoup

def scrape_flights(origin, destination, start_date, end_date):
    """
    Scrape flight data for the given parameters from Skyscanner (demo only).
    Returns: list of dicts with flight info.
    Note: Real scraping may require Selenium and anti-bot handling.
    """
    # This is a placeholder. Skyscanner and similar sites use heavy JavaScript and anti-bot measures.
    # For demo, return empty or mock data.
    # Example of what real data might look like:
    return [
        {
            'route': f'{origin}-{destination or "MEL"}',
            'date_time': start_date,
            'airline': 'Demo Airline',
            'price': 199,
            'status': 'scheduled'
        }
    ] 