requirements.txt
flask
requests
beautifulsoup4
selenium
pandas
numpy
python-dotenv 